pull-species/requirements.txt
pytest
pip-tools
biopython
python-dotenv